food_supply_chain/backend/requirements.txt
fastapi
uvicorn
web3
python-dotenv
pydantic
aiofiles
requests
ipfshttpclient
